Tank関連の不要なファイル削除
