+++ server/src/game/model/character/character.ts
@@ -4,7 +4,6 @@ import { ObjectUtil } from '../../util/object.util';
 import { GameObject } from '../gameObject/gameObject';
 import { GenericObstacle } from '../obstacle/generic/genericObstacle';
 import { GenericLinkedList } from '../../../linkedList/generic/genericLinkedList';
-import { Obstacle } from '../obstacle/interface/obstacle.interface';
 
 export class Character extends GameObject {
   static WIDTH = 38.4;
@@ -23,13 +22,11 @@ export class Character extends GameObject {
     Character.WIDTH,
     Character.HEIGHT
   );
-  //   private shouldUpdate = true
 
   // コンストラクタ
   constructor(
     public userName: string,
     spriteKey: string,
-    // obstacleSet: Set<GenericObstacle>
     obstacleList: GenericLinkedList<GenericObstacle>
   ) {
     super(0.0, 0.0, Character.WIDTH, Character.HEIGHT, spriteKey);
@@ -90,27 +87,6 @@ export class Character extends GameObject {
   }
 
   //初期位置に配置するメソッド
-  // setInitialPosition(obstacleSet: Set<GenericObstacle>) {
-  //   // 初期位置
-  //   this.setPosition(
-  //     Math.random() *
-  //       (CommonConfig.STAGE_WIDTH - this.getWidth),
-  //     Math.random() *
-  //       (CommonConfig.STAGE_HEIGHT - this.getHeight)
-  //   );
-
-  //   // 障害物にぶつからない初期位置の算出
-  //   do {
-  //     this.setPosition(
-  //       this.rectField.left +
-  //         Math.random() *
-  //           (this.rectField.right - this.rectField.left),
-  //       this.rectField.bottom +
-  //         Math.random() *
-  //           (this.rectField.top - this.rectField.bottom)
-  //     );
-  //   } while (this.overlapObstacles(obstacleSet));
-  // }
   setInitialPosition(obstacleList: GenericLinkedList<GenericObstacle>) {
     // 初期位置
     this.setPosition(
@@ -136,6 +112,7 @@ export class Character extends GameObject {
     );
   }
 
+  //アニメーションを設定するメソッド
   protected animWalkUp() {
     this.animation = `${this.spriteKey}-up`;
   }
