+++ server/src/game/model/player/player.ts
@@ -1,7 +1,7 @@
 import { GenericLinkedList } from '../../../linkedList/generic/genericLinkedList';
 import { Movement } from '../../types/movement.type';
 import { ObjectUtil } from '../../util/object.util';
-import { OverlapTester } from '../../util/overlapTester';
+import { OverlapUtil } from '../../util/overlap.util';
 import { Bomb } from '../bomb';
 import { Character } from '../character/character';
 import { GenericObstacle } from '../obstacle/generic/genericObstacle';
@@ -90,13 +90,13 @@ export class Player extends Character {
 
     //衝突判定
     let collision = false;
-    //補正値
+    //移動補正値
     let correction = {
       x: 0,
       y: 0,
     };
     if (
-      !OverlapTester.pointInRect(this.rectField, {
+      !OverlapUtil.pointInRect(this.rectField, {
         x: this.getPosition.x,
         y: this.getPosition.y,
       })
@@ -110,40 +110,8 @@ export class Player extends Character {
         // 障害物に当たった。
         collision = true;
 
-        //補正値を計算
+        //移動補正値を計算
         ObjectUtil.calCorrection(obstacleNode, this, correction);
-
-        // const obstacle = obstacleNode.data;
-        // if (
-        //   this.getPosition.x >= obstacle.getPosition.x &&
-        //   this.getPosition.y <= obstacle.getPosition.y
-        // ) {
-        //   const diffX = this.getPosition.x - obstacle.getPosition.x;
-        //   const diffY = obstacle.getPosition.y - this.getPosition.y;
-        //   if (diffX >= diffY) {
-        //     if (
-        //       diffY >=
-        //       ((obstacle.getHeight + this.getHeight) / 2) * (2 / 5)
-        //     ) {
-        //       //隣の障害物
-        //       let nextObstacle = obstacleNode.prev?.data;
-        //       if (nextObstacle) {
-        //         if (
-        //           !OverlapTester.overlapRects(
-        //             nextObstacle.rectBound,
-        //             this.rectBound
-        //           )
-        //         ) {
-        //           correction.y =
-        //             -1 *
-        //             (obstacle.getHeight / 2 +
-        //               this.getHeight / 2 -
-        //               (obstacle.getPosition.y - this.getPosition.y));
-        //         }
-        //       }
-        //     }
-        //   }
-        // }
       }
     }
     if (collision) {
