+++ server/src/linkedList/generic/genericAbstractList.ts
@@ -2,17 +2,7 @@
 export abstract class GenericAbstractList<E>
   
 {
-  private initialList: E[] = [];
-
   constructor() {}
-
-  //   constructor(arr: E[]) {
-  //     this.initialList = arr;
-  //   }
-
-  public getOriginalList(): E[] {
-    return this.initialList;
-  }
 
   public abstract addAt(index: number, element: E): void;
   public abstract removeAt(index: number): E | null;
