+++ server/src/socket/ioGame.ts
@@ -56,53 +56,6 @@ export default class IoGame {
           socket.roomId
         ].gameManager.game.stage.movePlayer(socket.clientId, movement);
       });
-      // socket.on('movePlayer', (movement: Movement) => {
-      //   if (!socket.roomId || !socket.clientId) return;
-
-      //   const playerSet =
-      //     this.roomManager.roomMap[socket.roomId]
-      //       .gameManager.game.stage.playerSet;
-
-      //   let foundPlayer = null;
-      //   playerSet.forEach((player) => {
-      //     if (player.clientId === socket.clientId) {
-      //       foundPlayer = player;
-      //       if (foundPlayer.getLife === 0) return;
-      //     }
-      //   });
-      //   if (!foundPlayer) return;
-
-      //   this.roomManager.roomMap[
-      //     socket.roomId
-      //   ].gameManager.game.stage.movePlayer(
-      //     socket.clientId,
-      //     movement
-      //   );
-      // });
-
-      // socket.on('moveTank', (objMovement: any) => {
-      //   if (!socket.roomId || !socket.clientId) return;
-
-      //   const tankSet =
-      //     this.roomManager.roomMap[socket.roomId]
-      //       .gameManager.game.stage.tankSet;
-      //   let foundTank = null;
-      //   tankSet.forEach((tank) => {
-      //     if (tank.clientId === socket.clientId) {
-      //       foundTank = tank;
-      //       if (foundTank.iLife === 0) return;
-      //     }
-      //   });
-      //   if (!foundTank) return;
-
-      //   // console.log(objMovement);
-      //   this.roomManager.roomMap[
-      //     socket.roomId
-      //   ].gameManager.game.stage.moveTank(
-      //     socket.clientId,
-      //     objMovement
-      //   );
-      // });
 
       //爆弾を設置するリクエストを受け取った時の処理
       socket.on('putBomb', () => {
@@ -130,57 +83,6 @@ export default class IoGame {
         //   socket.clientId
         // );
       });
-      // socket.on('putBomb', () => {
-      //   if (!socket.roomId || !socket.clientId) return;
-
-      //   console.log('⚠ 爆弾設置！！！');
-
-      //   const playerSet =
-      //     this.roomManager.roomMap[socket.roomId].gameManager.game.stage
-      //       .playerSet;
-      //   let foundPlayer = null;
-      //   playerSet.forEach((player) => {
-      //     if (player.clientId === socket.clientId) {
-      //       foundPlayer = player;
-      //       if (foundPlayer.getLife === 0) return;
-      //     }
-      //   });
-      //   if (!foundPlayer) return;
-
-      //   // 爆弾を設置
-      //   //createBombメソッドを後で実装する
-      //   // this.roomManager.roomMap[
-      //   //   socket.roomId
-      //   // ].gameManager.game.stage.createBomb(
-      //   //   socket.clientId
-      //   // );
-      // });
-
-      // ショット時の処理の指定
-      // クライアント側のキー入力時の「socket.emit( 'shoot' );」に対する処理
-      // socket.on('shoot', () => {
-      //   if (!socket.roomId || !socket.clientId) return;
-
-      //   const tankSet =
-      //     this.roomManager.roomMap[socket.roomId]
-      //       .gameManager.game.stage.tankSet;
-      //   let foundTank = null;
-      //   tankSet.forEach((tank) => {
-      //     if (tank.clientId === socket.clientId) {
-      //       foundTank = tank;
-      //       if (foundTank.iLife === 0) return;
-      //     }
-      //   });
-      //   if (!foundTank) return;
-
-      //   // ショット
-      //   this.roomManager.roomMap[
-      //     socket.roomId
-      //   ].gameManager.game.stage.createBullet(
-      //     socket.clientId,
-      //     0
-      //   );
-      // });
 
       //接続が切れたとき
       socket.on('disconnect', () => {
